First commit
Added a simple library to calculate the pagerank of a graph. The
PageRank is a sorting algorithm[1] initially used at Google to sort
pages.

It includes a ready-to-use binary as an example. This binary works with
the wikilinks[2] csv file.

```
gzcat eswiki.wikilink_graph.2018-03-01.csv.gz| cargo run --release wikilink
```

[1] http://ilpubs.stanford.edu:8090/422/1/1999-66.pdf
[2] https://zenodo.org/record/2539424